Use citation-styles for citations
