--- packages/11ty/_includes/components/citation/chicago/publishers.js
@@ -1,14 +1,0 @@
-/**
- * @param  {Object} context
- */
-module.exports = function(eleventyConfig) {
-  const { publisher: publishers } = eleventyConfig.globalData.publication
-
-  return function (params) {
-    if (!publishers || !publishers.length) return
-
-    return publishers
-      .map(({ location, name }) => [location, name].join(': '))
-      .join('; ')
-  }
-}
