--- packages/11ty/_includes/components/citation/pub-year.js
@@ -1,6 +1,0 @@
-module.exports = function(eleventyConfig) {
-  const pubDate = eleventyConfig.globalData.publication.pub_date
-  return function (params) {
-    return new Date(pubDate).getFullYear()
-  }
-}
