+++ packages/11ty/_includes/components/citation/publication.js
