+++ packages/11ty/_includes/components/index.js
@@ -8,22 +8,10 @@ module.exports = {
   abstract: require('./abstract.js'),
   analytics: require('./analytics.js'),
   citation: require('./citation/index.js'),
-  citationStylesLib: require('./citation/citation-styles-lib/index.js'),
-  citationStylesLibPage: require('./citation/citation-styles-lib/page.js'),
-  citationStylesLibPublication: require('./citation/citation-styles-lib/publication.js'),
-  citeContributors: require('./citation/contributors.js'),
   citeName: require('./citation/name.js'),
+  citePage: require('./citation/page.js'),
+  citePublication: require('./citation/publication.js'),
   citePublicationSeries: require('./citation/publication-series.js'),
-  chicagoPage: require('./citation/chicago/page/index.js'),
-  chicagoPagePublicationContributors: require ('./citation/chicago/page/publication-contributors.js'),
-  chicagoPublication: require ('./citation/chicago/publication/index.js'),
-  chicagoPublicationContributors: require ('./citation/chicago/publication/contributors.js'),
-  chicagoPublishers: require('./citation/chicago/publishers.js'),
-  MLAPage: require('./citation/mla/page/index.js'),
-  MLAPublication: require('./citation/mla/publication/index.js'),
-  MLAPublicationContributors: require('./citation/mla/publication/contributors.js'),
-  MLAPagePublicationContributors: require('./citation/mla/page/publication-contributors.js'),
-  MLAPublishers: require('./citation/mla/publishers.js'),
   contributorBio: require('./contributor/bio.js'),
   copyright: require('./copyright/index.js'),
   copyrightLicensing: require('./copyright/licensing.js'),
@@ -51,7 +39,6 @@ module.exports = {
   pageHeader: require('./page-header.js'),
   pageTitle: require('./page-title.js'),
   pdfInfo: require('./pdf/info.js'),
-  pubYear: require('./citation/pub-year.js'),
   scripts: require('./scripts.js'),
   siteTitle: require('./site-title.js'),
   tableOfContentsGridItem: require('./table-of-contents/grid-item.js'),
