--- packages/11ty/content/citations.md
@@ -1,43 +1,0 @@
----
-title: Quire Citations
-layout: page
-contributor:
-  - first_name: Hannah
-    last_name: Balenda
----
-
-### Chicago Manual of Style
-
-#### Publication
-
-`citation context='publication', page=pageData, type='chicago'`
-{% citation context='publication', page=pageData, type='chicago' %}
-
-`citationStylesLib context='publication', type='chicago'`
-{% citationStylesLib context='publication', type='chicago' %}
-
-#### Page
-
-`citation context='page', page=pageData, type='chicago'`
-{% citation context='page', page=pageData, type='chicago' %}
-
-`citationStylesLib context='page', page=pageData, type='chicago'`
-{% citationStylesLib context='page', page=pageData, type='chicago' %}
-
-### MLA
-
-#### Publication
-
-`citation context='publication', page=pageData, type='mla'`
-{% citation context='publication', page=pageData, type='mla' %}
-
-`citationStylesLib context='publication', type='mla'`
-{% citationStylesLib context='publication', type='mla' %}
-
-#### Page
-
-`citation context='page', page=pageData, type='mla'`
-{% citation context='page', page=pageData, type='mla' %}
-
-`citationStylesLib context='page', page=pageData, type='mla'`
-{% citationStylesLib context='page', page=pageData, type='mla' %}
